# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NewsHub API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Article {
  id: number;
  title: string;
  /** @nullable */
  source_url?: string | null;
  /** @nullable */
  image_url?: string | null;
  /** @nullable */
  summary?: string | null;
  /** @nullable */
  content?: string | null;
  /** @nullable */
  category?: string | null;
  /**
   * AI-generated industry commentary preview
   * @nullable
   */
  ai_comment?: string | null;
  /**
   * AI-generated one-liner for sharing to WeChat/social media
   * @nullable
   */
  share_text?: string | null;
  created_at: string;
  /** Estimated reading time in minutes */
  reading_time: number;
  /** Trending score 0-100 */
  hot_score: number;
}

export interface ArticleList {
  articles: Article[];
  total: number;
}

export interface Comment {
  id: number;
  article_id: number;
  content: string;
  commenter_name: string;
  created_at: string;
}

export interface AdjacentArticleRef {
  id: number;
  title: string;
}

export interface AdjacentArticles {
  prev?: AdjacentArticleRef | null;
  next?: AdjacentArticleRef | null;
}

export interface TriggerResult {
  success: boolean;
  message: string;
}

export interface StatsSummary {
  total_articles: number;
  total_comments: number;
  articles_today: number;
  /** @nullable */
  top_category: string | null;
  total_subscribers: number;
  subscribers_today: number;
}

export interface CategoryCount {
  category: string;
  count: number;
}

export interface TopArticle {
  id: number;
  title: string;
  /** @nullable */
  category?: string | null;
  views: number;
  created_at?: string;
}

export interface DailyViews {
  /** YYYY-MM-DD */
  date: string;
  views: number;
}

export type ListArticlesParams = {
  category?: string;
  limit?: number;
  offset?: number;
  search?: string;
};

export type ListCommentsParams = {
  article_id: number;
};

export type GetDailyViewsParams = {
  days?: number;
};