# lib/api-zod/src/generated/types/adjacentArticleRef.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NewsHub API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AdjacentArticleRef {
  id: number;
  title: string;
}