# lib/api-zod/src/generated/types/comment.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NewsHub API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Comment {
  id: number;
  article_id: number;
  content: string;
  commenter_name: string;
  created_at: Date;
}