# lib/api-zod/src/generated/types/statsSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NewsHub API specification
 * OpenAPI spec version: 0.1.0
 */

export interface StatsSummary {
  total_articles: number;
  total_comments: number;
  articles_today: number;
  /** @nullable */
  top_category: string | null;
  total_subscribers: number;
  subscribers_today: number;
}